(late) spring cleaning: docs and syntax lists
I moved things around and added several tables of contents to accommodate more
collection-level modules. The evaluator and example definer macros have been
cleaned up as well.

The old algebraic/syntax module is now at algebraic/private. The new
algebraic/syntax is a monad for context-preserving syntax transformations on
syntax lists. The code is pretty slick. I still need to document it.

I also added the `twice` function to the prelude, along with contracts on its
exports.

The new algebraic/syntax/list module uses the algebraic/syntax monad to
implement a portion of the "pairs and lists" functions for syntax objects. It
was good practice and the new code is cleaner.